git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk@1370 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ bogofilter/base64.h
@@ -12,6 +12,6 @@ NAME:
 
 #include "common.h"
 
-int base64_decode(byte *buff, size_t size);
+size_t base64_decode(byte *buff, size_t size);
 
 #endif	/* HAVE_BASE64_H */
